+++ src/schedulers/asap.jl
@@ -15,9 +15,9 @@ struct AsapScheduler <: Scheduler end
 Base.show(io::IO, ::AsapScheduler) = print(io, "AsapScheduler()")
 Base.similar(::AsapScheduler)      = AsapScheduler()
 
-subscribe!(::AsapScheduler, subscribable, actor) = on_subscribe!(subscribable, actor)
-unsubscribe!(::AsapScheduler, subscription)      = on_unsubscribe!(subscription)
+@inline subscribe!(::AsapScheduler, subscribable, actor) = on_subscribe!(subscribable, actor)
+@inline unsubscribe!(::AsapScheduler, subscription)      = on_unsubscribe!(subscription)
 
-next!(::AsapScheduler, actor, value) = on_next!(actor, value)
-error!(::AsapScheduler, actor, err)  = on_error!(actor, err)
-complete!(::AsapScheduler, actor)    = on_complete!(actor)
+@inline next!(::AsapScheduler, actor, value) = on_next!(actor, value)
+@inline error!(::AsapScheduler, actor, err)  = on_error!(actor, err)
+@inline complete!(::AsapScheduler, actor)    = on_complete!(actor)